@@ -1,6 +1,6 @@
 module github.com/lord-server/panorama
 
-go 1.20
+go 1.21
 
 require (
 	github.com/BurntSushi/toml v1.2.1
